Fix updater

+++ includes.sh
@@ -27,8 +27,13 @@ TZPREFS="$alfred_workflow_data"
 CONFIG_EXTRA="$TZPREFS/configExtra"
 TIMEZONE_PATH="$(getPreference 'TIMEZONE_PATH' "$TZPREFS" )"
 
+echo "alfred_workflow_data: $alfred_workflow_data" >> /tmp/alfred.txt
+mkdir "$alfred_workflow_data"
+echo "$alfred_workflow_data directory created" >> /tmp/alfred.txt
+
 #Load path to the user's timezones.txt file.
 timezone_file="$TIMEZONE_PATH/timezones.txt"
+echo "1 timezone_file: $timezone_file" >> /tmp/alfred.txt
 
 # echo "$TIMEZONE_PATH" >> /tmp/bzz
 # ls -l "$TIMEZONE_PATH" >> /tmp/bzz
@@ -39,17 +44,29 @@ shopt -s expand_aliases
 #Case-insensitive matching
 shopt -s nocasematch
 
+if [ ! -e "default_timezones.txt" ]; then
+  echo "default_timezones.txt" >> /tmp/alfred.txt
+  exit 1
+fi
+
 #Does the file actually exist?
 if [ ! -e "$timezone_file" ]; then
+  echo "timezone_file.txt does not exit. Creating it" >> /tmp/alfred.txt
 	#If not, recreate it from defaults
 	cp default_timezones.txt "$timezone_file"
 fi
 
-if ! grep 'Version2.0' "$timezone_file" > /dev/null
-then
+if [ ! -e "$timezone_file" ]; then
+  echo "timezone_file.txt still does not exit" >> /tmp/alfred.txt
+  exit 1
+fi
+
+if ! grep 'Version2.0' "$timezone_file" > /dev/null; then
+  echo "timezone_file.txt already exits. Overwriting it" >> /tmp/alfred.txt
 	cp default_timezones.txt "$timezone_file"
 fi
 
+echo "2 timezone_file: $timezone_file" >> /tmp/alfred.txt
 
 # Create an empty file (extra configuration) if it does not exist
 
@@ -57,10 +74,8 @@ if [[ ! -e "${CONFIG_EXTRA}"  ]]; then
 	touch "${CONFIG_EXTRA}"
 fi
 
-#
-# Preferences section
-#
-#
+##
+## Preferences section
 
 TIME_FORMAT=$(getPreference "TIME_FORMAT" "Both" )
 SORTING=$(getPreference "SORTING" "y" )
